+++ app/utils/generalUtils.ts
@@ -1,5 +1,6 @@
 import { useMatches } from "@remix-run/react"
 import { useMemo } from "react"
+import { isNonPositive } from "./numberUtils"
 
 const DEFAULT_REDIRECT = "/"
 
@@ -44,4 +45,15 @@ export function useMatchesData(
 
 export function validateEmail(email: unknown): email is string {
   return typeof email === "string" && email.length > 3 && email.includes("@")
+}
+
+export function parsePageNumberOrError(page: string): number {
+  const pageNumber = parseInt(page)
+  if (isNaN(pageNumber)) {
+    throw new Error(`Page Number ("${page}") must be a number`)
+  } else if (isNonPositive(pageNumber)) {
+    throw new Error(`Page Number ("${page}") must be greater than zero`)
+  }
+
+  return pageNumber
 }
